#1199.  Fixing msvc/64 build breakers (more size/long/uintXX_t conflicts).

+++ trunk/src/webpimage.cpp
@@ -373,14 +373,14 @@ namespace Exiv2 {
             if (outIo.write((const byte*)header.data(), 4) != 4)
                 throw Error(21);
 
-            us2Data(data, blob.size()+8, bigEndian);
+            us2Data(data, (uint16_t) blob.size()+8, bigEndian);
             byte exifHeader[] = { 0xFF, 0xE1, 0x00, 0x00,
                                   0x45, 0x78, 0x69, 0x66, 0x00, 0x00 };
             exifHeader[2] = data[0];
             exifHeader[3] = data[1];
             std::string rawExif =   std::string((char*)exifHeader, 10)
                                   + std::string((const char*)&blob[0], blob.size());
-            ul2Data(data, rawExif.size(), littleEndian);
+            ul2Data(data, (uint32_t) rawExif.size(), littleEndian);
             if (outIo.write(data, 4) != 4) throw Error(21);
             if (outIo.write((const byte*)rawExif.data(), static_cast<long>(rawExif.size())) != (long)rawExif.size())
             {
@@ -391,7 +391,7 @@ namespace Exiv2 {
         if (has_xmp) {
             std::string header = "XMP ";
             if (outIo.write((const byte*)header.data(), TAG_SIZE) != TAG_SIZE) throw Error(21);
-            ul2Data(data, xmpData.size(), littleEndian);
+            ul2Data(data, (uint32_t) xmpData.size(), littleEndian);
             if (outIo.write(data, 4) != 4) throw Error(21);
             if (outIo.write((const byte*)xmpData.data(), static_cast<long>(xmpData.size())) != (long)xmpData.size()) {
                 throw Error(21);
